LPS-198978 Extract service from KeyStoreCredentialResolver to KeyStoreLocalEntityManager, remove configurationPid from KeyStoreCredentialResolver since it doesn't have any usage, create KeyStoreUtil, apply to usages and fix unit tests
